# sample/Demo.Api.Generated/Resources/ApiSpecification.yaml
openapi: 3.0.0
info:
  title: Demo Api
  description: Demo Api - SingleFileVersion
  version: '1.0'
  contact:
    name: atc-net A/S
servers:
  - url: /api/v1
    description: Api version 1.0
paths:
...
  '/accounts/{accountId}/name':
    put:
      summary: Update name of account
      description: Update name of account
      operationId: updateAccountName
      responses:
        '200':
          description: OK
    post:
      summary: Set name of account
      description: Set name of account
      operationId: setAccountName
      responses:
        '200':
          description: OK
    parameters:
      - name: accountId
        in: path
        required: true
        description: The accountId
        schema:
          type: string
          format: uuid
...